Sets everything up to start Travis' deployment to Sonatype (#6)
Prepares the project for GPG signing of generated artifacts and their
upload to Sonatype.

Produced a CONTRIBUTING.md file with information on how to perform this
locally.

Adds hotfix branches to the `nonQualifierBranches` jgitver property so that in those branches the version does not get the name of the branch (i.e. 1.0.3-mybranch-SNAPSHOT).