Refactor generic test
This commit aims to refactor one of the tests on the AbstractProviderModelBaseTest.
The createOneModelAndEvaluateInMultipleThreadsTest test had some logic not as obvious as
it could be (IMHO - let me know if you disagree with this opinion). While we wanted to
check the number of occurrences of each classDistrbution result (assuming
each instance would produce a different output), the code was looking at
the minimum value found. This changes tries to make more obvious the goal of
that assertion.

Also added a toString method to the MockInstance, useful for test debug.