Add OpenML provider archetype
This commit adds a new module holding a maven archetype
to easily bootstrap a new OpenML provider.

The generated code does not try to be the most safe possible code
(for instance, when adding a new model, it needs to be manually added
to the provider list, a process that could be automated), but instead
the easiest entry point for someone new to the OpenML project.

Points that still could be improved:

- Javadoc
- tests (is it possible to have automated tests on maven archetypes?)

Work based on:

- https://dzone.com/articles/creating-a-custom-maven-archetype
- https://www.swtestacademy.com/maven-archetype/
- https://www.baeldung.com/maven-archetype
- https://medium.com/@swhp/invent-your-own-maven-archetype-7c882e5b44d1

How was it tested:

- `mvn install on the archetype sub module`
- on a new directory outside this git repo: `mvn archetype:generate -DarchetypeGroupId=com.feedzai -DarchetypeArtifactId=openml-provider-archetype -DarchetypeVersion=0.1.1-master-SNAPSHOT` (version may change due to jgitver)
- compile generated project