Make DatasetSchema target variables optional
Summary:
This commit changes dataset schemas (DatasetSchema) such that target variables are no longer needed. This was a limitation for Machine Learning Models, as certain
algorithms (mostly in unsupervised learning) have no notion of target variable, as they do not interpret class information.

Thus, from this commit, it is now possible to implement providers with algorithms of, for instance, unsupervised learning.