openml-utils: Update the documentation of the method AbstractProviderModelBaseTest#classifyIndexOfMaxScoresValue()
The documentation before was specifying the class as it was used only for the provider DataRobot.
However, it is an abstract test, hence the documentation should be abstract as well.
Therefore, it documents the method #classify() should return a number which is the index of the list's maximum value obtained from the method #getClassDistribution().